requirements.txt
imapclient
pyzmail36
requests
pandas
openpyxl
PyPDF2
